parser: parse headers containing invalid characters or codings
If there is a non-ascii character in a header, parsing fails,
even on Py27.

This has huge Py2/Py3 complexities. The Py3 email package has tools
to handle this - we just need to use them. Py2, on the other hand,
needs a lot of hand-holding, as explained in the comments.

Additionally, support headers that claim an encoding, but fail to
decode with that encoding.

This is handy for mails with malformed headers containing weird
bytes.

Reported-by: Thomas Monjalon <thomas.monjalon@6wind.com>
Suggested-by: Stephen Finucane <stephenfinucane@hotmail.com>
Signed-off-by: Daniel Axtens <dja@axtens.net>

---

Many thanks to Thomas for his help debugging this, and to Stephen
for insights leading to a much better patch - twice!

This should probably go to a stable branch too. We'll need to start
some discussion about how to handle bug fixes for people not running
git mainline (like ozlabs.org and kernel.org).

---

v2: - refactor to generalise header handling. Thanks Stephen.
    - support headers that claim an encoding but don't decode
      This actually leads to slightly simpler patch.