parsemail: Convert to a management command
Management comands allow applications to register their own actions with
'manage.py'. This provides some advantages, like automatically
configuring Django (removing the need for 'django.setup' calls) and
removing the need to set the PYTHON_PATH. The 'parsemail' script is a
natural fit for this type of application. Migrate 'parsemail' to a
management command.

This includes some extensive work on logging configuration, as logging
is moved from code into settings. In addition, it removes a lot of the
customizable logging previously introduced in the parsemail command, in
favour of modifications to the settings files.

Originally-by: Stephen Finucane <stephenfinucane@hotmail.com>
Partial-bug: #17
Closes-bug: #15
[dja: significant changes:
 - fix merge conflicts with my changes to parsemail.sh
 - py3 binary file email parsing
 - make stdin tests close and re-open stdin]
Signed-off-by: Daniel Axtens <dja@axtens.net>

---

Stephen: the changes here are significant enough that I thought it
might be inappropriate to keep your signed-off-by until you've had
a chance to review them. If you're happy with them I'm more than
happy for you to add your sign-off when you merge it.