Burninate CoverLetter
A patch must have a diff or a pull_url.
If a Submission does not have a diff or a pull_url, it must be
a CoverLetter.

This is enough to get rid of the CoverLetter model. We do have some
work to do around the interactions with Series, which is currently
super ugly, but it's good enough for a proof of concept.

Must preceed series migration due to that interaction! (or we have to
tweak the backref in series which would be much worse.)

Signed-off-by: Daniel Axtens <dja@axtens.net>